Add handling for redirects for logged in users
If a user is logged in, but a page is configured to "hide at login"
We now check "redirectError403LoggedInUserTo" TS config where you
can configure e.g. user profile page uid.
So a logged in user will be redirected to his profile after accessing
a page which is hidden at login.